requirements.txt
# Slack
slack-bolt>=1.18.0
slack-sdk>=3.26.0

# Web / API
fastapi>=0.110.0
uvicorn>=0.27.0
httpx>=0.27.0

# LLM
openai>=1.30.0

# Data
pydantic>=2.0.0
python-dotenv>=1.0.0

# Utils
loguru>=0.7.0
